Move the HttpOperation class out of RPCService
