+++ modules/internal/src/main/scala/higherkindness/mu/rpc/internal/service/makro/RPCServiceModel.scala
@@ -3,6 +3,7 @@ package higherkindness.mu.rpc.internal.service.makro
 import higherkindness.mu.rpc.protocol._
 import scala.reflect.macros.blackbox.Context
 
+// $COVERAGE-OFF$
 class RPCServiceModel[C <: Context](val c: C) {
   import c.universe._
 
@@ -338,123 +339,7 @@ class RPCServiceModel[C <: Context](val c: C) {
         case _                                                     => false
       }
 
-    case class HttpOperation(operation: Operation) {
-
-      import operation._
-
-      val uri = name.toString
-
-      val method: TermName = request match {
-        case _: EmptyTpe => TermName("GET")
-        case _           => TermName("POST")
-      }
-
-      val requestTypology: Tree = request match {
-        case _: UnaryTpe =>
-          q"val request = _root_.org.http4s.Request[F](_root_.org.http4s.Method.$method, uri / ${uri
-            .replace("\"", "")}).withEntity(req.asJson)"
-        case _: Fs2StreamTpe =>
-          q"val request = _root_.org.http4s.Request[F](_root_.org.http4s.Method.$method, uri / ${uri
-            .replace("\"", "")}).withEntity(req.map(_.asJson))"
-        case _ =>
-          q"val request = _root_.org.http4s.Request[F](_root_.org.http4s.Method.$method, uri / ${uri
-            .replace("\"", "")})"
-      }
-
-      val executionClient: Tree = response match {
-        case _: Fs2StreamTpe =>
-          q"_root_.cats.Applicative[F].pure(client.stream(request).flatMap(_.asStream[${response.messageType}]))"
-        case _ =>
-          q"""client.expectOr[${response.messageType}](request)(handleResponseError)(_root_.org.http4s.circe.jsonOf[F, ${response.messageType}])"""
-      }
-
-      def toRequestTree: Tree = request match {
-        case _: EmptyTpe =>
-          q"""def $name(client: _root_.org.http4s.client.Client[F])(
-               implicit responseDecoder: _root_.io.circe.Decoder[${response.messageType}]): ${response.originalType} = {
-                 $requestTypology
-                 $executionClient
-               }"""
-        case _ =>
-          q"""def $name(req: ${request.originalType})(client: _root_.org.http4s.client.Client[F])(
-               implicit requestEncoder: _root_.io.circe.Encoder[${request.messageType}],
-               responseDecoder: _root_.io.circe.Decoder[${response.messageType}]
-            ): ${response.originalType} = {
-              $requestTypology
-              $executionClient
-            }"""
-      }
-
-      val routeTypology: Tree = (request, response) match {
-        // Stream -> Stream
-        case (_: Fs2StreamTpe, _: Fs2StreamTpe) =>
-          q"""val requests = msg.asStream[${operation.request.messageType}]
-              for {
-                respStream <- handler.${operation.name}(requests)
-                responses  <- _root_.org.http4s.Status.Ok.apply(respStream.asJsonEither)
-              } yield responses"""
-
-        // Stream -> Empty
-        case (_: Fs2StreamTpe, _: EmptyTpe) =>
-          q"""val requests = msg.asStream[${operation.request.messageType}]
-              _root_.org.http4s.Status.Ok.apply(handler.${operation.name}(requests).as(()))"""
-
-        // Stream -> Unary
-        case (_: Fs2StreamTpe, _: UnaryTpe) =>
-          q"""val requests = msg.asStream[${operation.request.messageType}]
-              _root_.org.http4s.Status.Ok.apply(handler.${operation.name}(requests).map(_.asJson))"""
-
-        // Empty -> Stream
-        case (_: EmptyTpe, _: Fs2StreamTpe) =>
-          q"""for {
-                respStream <- handler.${operation.name}(_root_.higherkindness.mu.rpc.protocol.Empty)
-                responses  <- _root_.org.http4s.Status.Ok.apply(respStream.asJsonEither)
-              } yield responses"""
-
-        // Empty -> Empty
-        case (_: EmptyTpe, _: EmptyTpe) =>
-          q"""_root_.org.http4s.Status.Ok.apply(handler.${operation.name}(_root_.higherkindness.mu.rpc.protocol.Empty).as(()))"""
-
-        // Empty -> Unary
-        case (_: EmptyTpe, _: UnaryTpe) =>
-          q"""_root_.org.http4s.Status.Ok.apply(handler.${operation.name}(_root_.higherkindness.mu.rpc.protocol.Empty).map(_.asJson))"""
-
-        // Unary -> Stream
-        case (_: UnaryTpe, _: Fs2StreamTpe) =>
-          q"""for {
-              request    <- msg.as[${operation.request.messageType}]
-              respStream <- handler.${operation.name}(request)
-              responses  <- _root_.org.http4s.Status.Ok.apply(respStream.asJsonEither)
-            } yield responses"""
-
-        // Unary -> Empty
-        case (_: UnaryTpe, _: EmptyTpe) =>
-          q"""for {
-              request  <- msg.as[${operation.request.messageType}]
-              response <- _root_.org.http4s.Status.Ok.apply(handler.${operation.name}(request).as(())).adaptErrors
-            } yield response"""
-
-        // Unary -> Unary
-        case _ =>
-          q"""for {
-              request  <- msg.as[${operation.request.messageType}]
-              response <- _root_.org.http4s.Status.Ok.apply(handler.${operation.name}(request).map(_.asJson)).adaptErrors
-            } yield response"""
-      }
-
-      val getPattern =
-        pq"_root_.org.http4s.Method.GET -> _root_.org.http4s.dsl.impl.Root / ${operation.name.toString}"
-      val postPattern =
-        pq"msg @ _root_.org.http4s.Method.POST -> _root_.org.http4s.dsl.impl.Root / ${operation.name.toString}"
-
-      def toRouteTree: Tree = request match {
-        case _: EmptyTpe => cq"$getPattern => $routeTypology"
-        case _           => cq"$postPattern => $routeTypology"
-      }
-
-    }
-
-    val operations: List[HttpOperation] = for {
+    val httpOperations: List[HttpOperation] = for {
       d <- rpcDefs.collect { case x if findAnnotation(x.mods, "http").isDefined => x }
       // TODO not sure what the following line is doing, as the result is not used. Is it needed?
       _      <- findAnnotation(d.mods, "http").collect({ case Apply(_, args) => args }).toList
@@ -472,7 +357,7 @@ class RPCServiceModel[C <: Context](val c: C) {
 
     val streamConstraints: List[Tree] = List(q"F: _root_.cats.effect.Sync[$F]")
 
-    val httpRequests = operations.map(_.toRequestTree)
+    val httpRequests = httpOperations.map(_.toRequestTree)
 
     val HttpClient      = TypeName("HttpClient")
     val httpClientClass = q"""
@@ -494,18 +379,18 @@ class RPCServiceModel[C <: Context](val c: C) {
       q"import _root_.io.circe.syntax._"
     )
 
-    val httpRoutesCases: Seq[Tree] = operations.map(_.toRouteTree)
+    val httpRoutesCases: Seq[Tree] = httpOperations.map(_.toRouteTree)
 
     val routesPF: Tree = q"{ case ..$httpRoutesCases }"
 
     val requestTypes: Set[String] =
-      operations
+      httpOperations
         .filterNot(_.operation.request.isEmpty)
         .map(_.operation.request.messageType.toString)
         .toSet
 
     val responseTypes: Set[String] =
-      operations
+      httpOperations
         .filterNot(_.operation.response.isEmpty)
         .map(_.operation.response.messageType.toString)
         .toSet
@@ -547,3 +432,4 @@ class RPCServiceModel[C <: Context](val c: C) {
   }
 
 }
+// $COVERAGE-ON$
